chore: extract overview/config from dashboard index (move-only)
