+++ src/routes/dashboard/index.ts
@@ -102,12 +102,14 @@ import { WebullAuth } from '../../infrastructure/webull/WebullAuth'
 import { WebullTokenClient } from '../../infrastructure/webull/WebullTokenClient'
 import { WebullTokenStateClient } from '../../trading/state/WebullTokenStateClient'
 import type { WebullTokenState } from '../../trading/state/WebullTokenStateDO'
-import { clampLimit, currencyOfSymbol, displaySymbol, esc, fmtJst, fmtNumber, fmtPct, fmtPctSigned, fmtPriceCcy, inactiveTooltip, isSymbolInactive, jsonPretty, messageOf, parseCursor, safeJsonScript, unavailable } from './shared'
+import { clampLimit, currencyOfSymbol, displaySymbol, esc, fmtPct, fmtPctSigned, fmtPriceCcy, inactiveTooltip, isSymbolInactive, jsonPretty, messageOf, parseCursor, safeJsonScript, unavailable } from './shared'
 import { type DashboardBindings, layout, loadKillSwitchState, renderLayout } from './layout'
-import { loadLatestStrategyPrices, pickFreshQuote, positionsBody } from './positions'
-import { type EquityRange, parseEquityRange, portfolioBody, renderPortfolioEquityChart, renderVixRegimeCell, safeLoadPortfolioSnapshots } from './portfolio'
+import { ALL_OVERVIEW_PANELS, type OverviewData, buyingPowerBadge, loadRecentFills, overviewBody, parseOverviewPanels } from './overview'
+import { loadLatestStrategyPrices, positionsBody } from './positions'
+import { parseEquityRange, portfolioBody, safeLoadPortfolioSnapshots } from './portfolio'
 import { tradesBody } from './trades'
-import { cronBody, cronDecisionJson, formatRealizedPnl, loadDecisionRows, renderDecisionLadder, renderDecisionTable } from './cron'
+import { configBody } from './config'
+import { cronBody, cronDecisionJson, loadDecisionRows, renderDecisionLadder, renderDecisionTable } from './cron'
 import { alertsBody, clampAlertLimit, parseAlertsQuery, parseEventTypeFilter, parseSeverityFilter } from './alerts'
 import { auditBody, clampAuditLimit, parseAuditDateFilter, trimQuery } from './audit'
 import { type ChartsBodyArgs, type ChartsBodyGrid, type ChartsBodySymbol, ECHARTS_CDN, type StrategyParamsSnapshot, type SymbolPolicySummary, computeZoomRange, parseChartsTab, parseIsoTimestamp, renderChartsSubnav, renderZoomPresetButtons } from './charts/shared'
@@ -115,6 +117,8 @@ import { MAX_CHART_DECISIONS, type SymbolChartData, type SymbolChartPoint, type 
 import { loadEquityCurve, renderOverviewTab } from './charts/equity'
 import { computePnlHistogram, computeTradeStats, loadDecisionBreakdown, loadTradePnls, renderQualityTab } from './charts/quality'
 export { safeJsonScript } from './shared'
+export { ALL_OVERVIEW_PANELS, OVERVIEW_PANEL_LABELS, parseOverviewPanels } from './overview'
+export type { OverviewPanel } from './overview'
 export { formatQuoteAsOf, pickFreshQuote } from './positions'
 export { parseEquityRange, renderLastRolledCell, renderVixRegimeCell } from './portfolio'
 export type { EquityRange } from './portfolio'
@@ -2316,520 +2320,6 @@ function brokerProbeBody(args: {
   }
 })();
 </script>`
-}
-
-// #dashboard-mf-layout: overview パネル定義。設定で ON/OFF (default 全表示)。
-export type OverviewPanel = 'kpi' | 'equity' | 'composition' | 'recent'
-export const ALL_OVERVIEW_PANELS: readonly OverviewPanel[] = ['kpi', 'equity', 'composition', 'recent']
-export const OVERVIEW_PANEL_LABELS: Record<OverviewPanel, string> = {
-  kpi: 'KPI カード (総資産 / 当日損益 / 建玉数 / エクスポージャー)',
-  equity: '資産推移チャート (期間タブ)',
-  composition: '資産構成 + 含み損益ランキング',
-  recent: '最近の約定 + VIX / リスク状態',
-}
-
-/** CSV を有効パネル集合へ。不正値は無視、空 (未設定/全部不正) は全表示。 */
-export function parseOverviewPanels(csv: string | null | undefined): Set<OverviewPanel> {
-  const set = new Set<OverviewPanel>()
-  for (const tok of (csv ?? '').split(',').map((s) => s.trim())) {
-    if ((ALL_OVERVIEW_PANELS as readonly string[]).includes(tok)) set.add(tok as OverviewPanel)
-  }
-  return set.size === 0 ? new Set(ALL_OVERVIEW_PANELS) : set
-}
-
-interface OverviewData {
-  panels: Set<OverviewPanel>
-  portfolio: {
-    dailyStartEquity: number
-    dailyRealizedPnl: number
-    openExposureUsd: number
-    openExposureJpy: number
-    tradingDisabledUntil: string | null
-    updatedAt: string
-  } | null
-  snapshots: PortfolioEquitySnapshotRow[]
-  range: EquityRange
-  positions: Array<{ sym: string; state: SymbolState | null; error: string | null }>
-  strategyPriceMap: Map<string, { price: number; asOf: string }>
-  recentTrades: Array<{
-    id: number
-    timestamp: string
-    symbol: string | null
-    side: string | null
-    filledQty: number | null
-    filledPrice: number | null
-    realizedPnl: number | null
-    brokerStatus: string | null
-  }>
-  vixRegime: VixRegime | null
-  dryRun: boolean
-  tradingEnabled: boolean
-  universe: SymbolUniverse
-}
-
-/** 開いている建玉 (qty != 0) を評価額・含み損益% 付きで抽出。 */
-interface OpenPositionView {
-  sym: string
-  qty: number
-  currency: SymbolCurrency
-  price: number | null
-  marketValue: number | null
-  pnlPct: number | null
-}
-
-/**
- * overview「最近の約定」用の直近 fill ロード。post_submit 行は side が null
- * (writer は pre_submit にしか入れない) なので client_order_id で pre_submit と
- * self-JOIN して side を引く (loadSymbolChart と同方針)。pre_submit が無い古い fill は
- * realized_pnl の有無から推測 (null=BUY / 非null=SELL)。
- */
-async function loadRecentFills(
-  db: D1Database,
-  limit: number,
-): Promise<OverviewData['recentTrades']> {
-  const result = await db
-    .prepare(
-      `SELECT ps.id AS id, ps.timestamp AS timestamp, ps.symbol AS symbol,
-         COALESCE(pre.side, CASE WHEN ps.realized_pnl IS NOT NULL THEN 'SELL' ELSE 'BUY' END) AS side,
-         ps.filled_qty AS filledQty, ps.filled_price AS filledPrice,
-         ps.realized_pnl AS realizedPnl, ps.broker_status AS brokerStatus
-       FROM trade_journal AS ps
-       LEFT JOIN trade_journal AS pre
-         ON pre.client_order_id = ps.client_order_id AND pre.trade_event_type = 'pre_submit'
-       WHERE ps.trade_event_type = 'post_submit' AND ps.filled_price IS NOT NULL
-       ORDER BY ps.id DESC
-       LIMIT ?`,
-    )
-    .bind(limit)
-    .all<{
-      id: number
-      timestamp: string
-      symbol: string | null
-      side: string | null
-      filledQty: number | null
-      filledPrice: number | null
-      realizedPnl: number | null
-      brokerStatus: string | null
-    }>()
-  return result.results ?? []
-}
-
-function collectOpenPositions(data: OverviewData): OpenPositionView[] {
-  const out: OpenPositionView[] = []
-  for (const r of data.positions) {
-    const pos = r.state?.position
-    if (!r.state || !pos || pos.qty === 0) continue
-    const webull = r.state.lastQuote
-      ? { price: r.state.lastQuote.price, source: r.state.lastQuote.source, asOf: r.state.lastQuote.asOf ?? r.state.lastQuote.fetchedAt }
-      : null
-    const yahoo = data.strategyPriceMap.get(r.state.symbol) ?? null
-    const quote = pickFreshQuote(webull, yahoo)
-    const price = quote?.price ?? null
-    const pnlPct = price !== null && pos.avgPrice > 0 ? ((price - pos.avgPrice) / pos.avgPrice) * 100 : null
-    out.push({
-      sym: r.state.symbol,
-      qty: pos.qty,
-      currency: data.universe.symbolCurrency[r.state.symbol] ?? 'USD',
-      price,
-      marketValue: price !== null ? pos.qty * price : null,
-      pnlPct,
-    })
-  }
-  return out
-}
-
-function kpiCard(label: string, value: string, sub?: string, subClass?: string): string {
-  const subHtml = sub ? `<div class="kpi-sub ${subClass ?? 'muted'}">${sub}</div>` : ''
-  return `<div class="kpi-card"><div class="kpi-label">${esc(label)}</div><div class="kpi-value">${value}</div>${subHtml}</div>`
-}
-
-function renderKpiPanel(data: OverviewData, open: OpenPositionView[]): string {
-  const p = data.portfolio
-  const dd = p && p.dailyStartEquity > 0 ? (p.dailyRealizedPnl / p.dailyStartEquity) * 100 : null
-  const pnlClass = p == null ? 'muted' : p.dailyRealizedPnl >= 0 ? 'ok' : 'err'
-  const cards = [
-    kpiCard('当日始値資産', p ? fmtNumber(p.dailyStartEquity, 2) : '—', '口座 dailyStartEquity'),
-    kpiCard(
-      '当日実現損益',
-      p ? `<span class="${pnlClass}">${fmtNumber(p.dailyRealizedPnl, 2)}</span>` : '—',
-      dd === null ? undefined : `DD ${fmtNumber(dd, 2)}%`,
-      dd === null ? 'muted' : dd >= 0 ? 'ok' : 'err',
-    ),
-    kpiCard('建玉数', String(open.length), '保有中の銘柄数'),
-    kpiCard(
-      'Open exposure',
-      p ? `${fmtNumber(p.openExposureUsd, 0)}<span class="muted" style="font-size:12px"> USD</span>` : '—',
-      p ? `${fmtNumber(p.openExposureJpy, 0)} JPY` : undefined,
-    ),
-  ].join('')
-  return `<div class="kpi-grid">${cards}</div>`
-}
-
-function renderCompositionPanel(open: OpenPositionView[]): string {
-  if (open.length === 0) {
-    return `<div class="panel"><div class="panel-title">資産構成 / 含み損益ランキング</div><p class="muted">保有中の建玉がありません。</p></div>`
-  }
-  // 通貨内シェアで構成比 bar を正規化 (USD/JPY を混ぜない)。
-  const sumByCcy: Record<string, number> = {}
-  for (const o of open) {
-    if (o.marketValue !== null) sumByCcy[o.currency] = (sumByCcy[o.currency] ?? 0) + Math.abs(o.marketValue)
-  }
-  const composition = [...open]
-    .sort((a, b) => (Math.abs(b.marketValue ?? 0)) - (Math.abs(a.marketValue ?? 0)))
-    .map((o) => {
-      const total = sumByCcy[o.currency] ?? 0
-      const share = o.marketValue !== null && total > 0 ? (Math.abs(o.marketValue) / total) * 100 : 0
-      const valueText = o.marketValue !== null ? `${fmtNumber(o.marketValue, 0)} ${o.currency}` : '—'
-      return `<div class="rank-row"><span>${esc(o.sym)} <span class="muted" style="font-size:11px">${fmtNumber(share, 1)}%</span></span><span>${valueText}</span></div>
-        <div class="bar-track"><div class="bar-fill" style="width:${share.toFixed(1)}%"></div></div>`
-    })
-    .join('')
-  // 含み損益% ランキング (up / down)。
-  const ranked = open.filter((o) => o.pnlPct !== null) as Array<OpenPositionView & { pnlPct: number }>
-  const gainers = [...ranked].filter((o) => o.pnlPct >= 0).sort((a, b) => b.pnlPct - a.pnlPct).slice(0, 5)
-  const losers = [...ranked].filter((o) => o.pnlPct < 0).sort((a, b) => a.pnlPct - b.pnlPct).slice(0, 5)
-  const rankRow = (o: OpenPositionView & { pnlPct: number }) =>
-    `<div class="rank-row"><span>${esc(o.sym)}</span><span class="${o.pnlPct >= 0 ? 'ok' : 'err'}">${fmtNumber(o.pnlPct, 2)}%</span></div>`
-  const rankCol = (title: string, items: Array<OpenPositionView & { pnlPct: number }>) =>
-    `<div><div class="muted" style="font-size:12px;margin-bottom:4px">${esc(title)}</div>${items.length ? items.map(rankRow).join('') : '<p class="muted">—</p>'}</div>`
-  return `<div class="panel">
-    <div class="panel-title">資産構成 / 含み損益ランキング</div>
-    <p class="muted" style="font-size:12px;margin-top:0">構成比は通貨内シェア。ランキングは含み損益% (現在値 vs 平均取得単価)。</p>
-    <div class="panel-row">
-      <div><div class="muted" style="font-size:12px;margin-bottom:4px">構成 (評価額)</div>${composition}</div>
-      <div class="panel-row" style="grid-template-columns:1fr 1fr">${rankCol('上昇', gainers)}${rankCol('下落', losers)}</div>
-    </div>
-  </div>`
-}
-
-function renderRecentPanel(data: OverviewData): string {
-  const trades = data.recentTrades
-    .map((t) => {
-      const sideClass = t.side === 'BUY' ? 'ok' : t.side === 'SELL' ? 'err' : 'muted'
-      const pnl = t.realizedPnl !== null ? formatRealizedPnl(t.realizedPnl) : '<span class="muted">—</span>'
-      return `<tr>
-        <td class="muted" style="font-size:12px">${esc(fmtJst(t.timestamp))}</td>
-        <td><strong>${esc(displaySymbol(t.symbol ?? '—', data.universe))}</strong></td>
-        <td class="${sideClass}">${esc(t.side ?? '—')}</td>
-        <td>${t.filledQty !== null ? esc(t.filledQty) : '—'}</td>
-        <td>${t.filledPrice !== null ? fmtNumber(t.filledPrice, 2) : '—'}</td>
-        <td>${pnl}</td>
-      </tr>`
-    })
-    .join('')
-  const recentTable = data.recentTrades.length
-    ? `<table><thead><tr><th>時刻</th><th>銘柄</th><th>売買</th><th>数量</th><th>約定値</th><th>実損益</th></tr></thead><tbody>${trades}</tbody></table>`
-    : '<p class="muted">約定履歴がありません。</p>'
-  const dryPill = data.dryRun
-    ? '<span class="pill dry">DRY-RUN</span>'
-    : '<span class="pill live">LIVE</span>'
-  const tradingPill = data.tradingEnabled
-    ? '<span class="pill on">取引 ON</span>'
-    : '<span class="pill off">取引 OFF</span>'
-  return `<div class="panel">
-    <div class="panel-title">最近の約定 / リスク状態</div>
-    <div class="panel-row">
-      <div>${recentTable}<div style="margin-top:8px"><a href="/dashboard/trades">約定履歴をすべて見る →</a></div></div>
-      <div>
-        <table><tbody>
-          <tr><th>実行モード</th><td>${dryPill} <span class="muted" style="font-size:11px">(D1 dry_run)</span></td></tr>
-          <tr><th>取引 (effective)</th><td>${tradingPill} <span class="muted" style="font-size:11px">(env override 反映後)</span></td></tr>
-          <tr><th>VIX レジーム</th><td>${renderVixRegimeCell(data.vixRegime)}</td></tr>
-        </tbody></table>
-      </div>
-    </div>
-  </div>`
-}
-
-/**
- * 口座買付余力バッジ (#415)。SSR はブロックせず、client-side で `/admin/buying-power`
- * を fetch して通貨別 buying_power を描画する (broker-probe と同じく CF Access cookie
- * 流用)。取得失敗は ⚠ 表示でページは壊さない。ホーム / 銘柄設定の両方で使う。
- */
-function buyingPowerBadge(): string {
-  return `<div id="buying-power-badge" class="panel" style="display:flex;align-items:center;gap:8px;font-size:13px;padding:10px 14px">
-    <strong>買付余力</strong> <span class="muted">読込中…</span>
-  </div>
-  <script>
-  (function () {
-    var el = document.getElementById('buying-power-badge');
-    if (!el) return;
-    fetch('/admin/buying-power', { credentials: 'same-origin' })
-      .then(function (r) { return r.ok ? r.json() : { status: 'unavailable', reason: 'http ' + r.status }; })
-      .then(function (d) {
-        if (!d || d.status !== 'ok') {
-          el.innerHTML = '<strong>買付余力</strong> <span style="color:#c22;font-weight:600">⚠ 取得不可</span>' +
-            ' <span class="muted" style="font-size:11px">' + ((d && d.reason) ? String(d.reason).slice(0, 80) : '') + '</span>';
-          return;
-        }
-        var parts = (d.byCurrency || []).map(function (a) {
-          var bp = Number(a.buyingPower);
-          var sym = a.currency === 'JPY' ? '¥' : (a.currency === 'USD' ? '$' : '');
-          var v = isFinite(bp) ? bp.toLocaleString('ja-JP', { maximumFractionDigits: a.currency === 'JPY' ? 0 : 2 }) : a.buyingPower;
-          var zero = isFinite(bp) && bp <= 0;
-          return '<span style="' + (zero ? 'color:#86868b' : 'font-weight:600') + '">' + a.currency + ' ' + sym + v + '</span>';
-        });
-        el.innerHTML = '<strong>買付余力</strong> ' + (parts.join(' &nbsp;/&nbsp; ') || '—') +
-          ' <span class="muted" style="font-size:11px">(口座 ' + (d.baseCurrency || '') + ' 総現金 ' + Number(d.totalCash).toLocaleString('ja-JP') + ')</span>';
-      })
-      .catch(function () {
-        el.innerHTML = '<strong>買付余力</strong> <span style="color:#c22;font-weight:600">⚠ 取得不可</span>';
-      });
-  })();
-  </script>`
-}
-
-function overviewBody(data: OverviewData): string {
-  const open = collectOpenPositions(data)
-  const sections: string[] = []
-  if (data.panels.has('kpi')) sections.push(renderKpiPanel(data, open))
-  if (data.panels.has('equity')) {
-    sections.push(
-      `<div class="panel">${renderPortfolioEquityChart(data.snapshots, data.range, '/dashboard')}</div>`,
-    )
-  }
-  if (data.panels.has('composition')) sections.push(renderCompositionPanel(open))
-  if (data.panels.has('recent')) sections.push(renderRecentPanel(data))
-  if (sections.length === 0) {
-    sections.push(
-      '<p class="muted">表示パネルが選択されていません。<a href="/dashboard/config">設定</a>でパネルを有効化してください。</p>',
-    )
-  }
-  return buyingPowerBadge() + sections.join('')
-}
-
-function configBody(
-  global: Awaited<ReturnType<typeof loadGlobalConfigFrom>>,
-  universe: Awaited<ReturnType<typeof loadSymbolUniverse>>,
-  overviewPanels: Set<OverviewPanel>,
-): string {
-  // #dashboard-mf-layout: overview パネル ON/OFF。POST → PRG redirect (#293 と同型)。
-  const panelForm = `<details open>
-    <summary>ダッシュボード overview パネル表示</summary>
-    <form method="post" action="/dashboard/config/overview-panels" style="margin:8px 0;display:flex;flex-direction:column;gap:6px;max-width:560px">
-      ${ALL_OVERVIEW_PANELS.map((k) => `<label style="font-size:13px"><input type="checkbox" name="panels" value="${k}"${overviewPanels.has(k) ? ' checked' : ''}/> ${esc(OVERVIEW_PANEL_LABELS[k])}</label>`).join('')}
-      <div><button type="submit" style="padding:4px 12px;font-size:13px;background:#06c;color:#fff;border:none;border-radius:6px;cursor:pointer">保存</button></div>
-    </form>
-    <p class="muted" style="font-size:12px"><code>/dashboard</code> の overview に表示するパネル。全て OFF にすると全表示に戻ります。</p>
-  </details>`
-  // 列名 (snake_case) は SQL での copy-paste 互換のため英字のまま残し、
-  // 日本語説明は別列に分離。これで `UPDATE global_config SET xxx = ...` が
-  // そのまま使える。
-  const globalRows = Object.entries(global as unknown as Record<string, unknown>)
-    .filter(([k]) => k !== 'source')
-    .map(([k, v]) => {
-      const camelKey = k.replace(/[A-Z]/g, (c) => '_' + c.toLowerCase())
-      // DB 列名の digit 前 underscore は列ごとに揺れがある
-      // (min_return_50d は有 / require_above_sma50 は無)。
-      // naive 版 → digit 前 underscore 版の順でフォールバック。
-      const camelKeyWithDigitUnderscore = camelKey.replace(/([a-z])(\d)/g, '$1_$2')
-      const meta =
-        CONFIG_KEY_META[camelKey] ??
-        CONFIG_KEY_META[camelKeyWithDigitUnderscore] ??
-        CONFIG_KEY_META[k]
-      const label = meta?.label ?? '—'
-      const detail = meta?.detail ?? '—'
-      return `<tr><th>${esc(camelKey)}</th><td>${esc(formatConfigValue(v))}</td><td class="muted">${esc(label)}</td><td class="muted" style="font-size:11px">${esc(detail)}</td></tr>`
-    })
-    .join('')
-  // active + inactive 両方を 1 つの table で表示。inactive 行は grayed-out 化し、
-  // 「状態」「メモ (notes)」列で disable 経緯が読める。cron 評価対象は active=1 のみ
-  // (allowedSymbols)、表示のみ全件出すのが今 PR の趣旨。
-  const allConfigSymbols = [...universe.allowedSymbols, ...universe.inactiveSymbols]
-  const symRows = allConfigSymbols
-    .map((sym) => {
-      const inactive = isSymbolInactive(sym, universe)
-      const rowClass = inactive ? ' class="symbol-disabled-row"' : ''
-      const symbolClass = inactive ? ' class="symbol-disabled"' : ''
-      const titleAttr = inactive ? ` title="${esc(inactiveTooltip(sym, universe))}"` : ''
-      const stateCell = inactive
-        ? '<span class="muted">inactive</span>'
-        : '<span class="ok">active</span>'
-      const noteText = universe.symbolNotes[sym] ?? null
-      const noteCell = noteText ? esc(noteText) : '<span class="muted">—</span>'
-      return `<tr${rowClass}>
-          <td><strong><span${symbolClass}${titleAttr}>${esc(displaySymbol(sym, universe))}</span></strong></td>
-          <td>${stateCell}</td>
-          <td>${esc(universe.symbolCurrency[sym] ?? '—')}</td>
-          <td>${universe.symbolMaxNotional[sym] != null ? esc(universe.symbolMaxNotional[sym]) : '<span class="muted">—</span>'}</td>
-          <td>${universe.inversePairs[sym] ? esc(universe.inversePairs[sym]) : '<span class="muted">—</span>'}</td>
-          <td>${noteCell}</td>
-        </tr>`
-    })
-    .join('')
-  return `${panelForm}
-  <details open>
-    <summary>グローバル設定 (global_config)</summary>
-    <table>
-      <thead><tr><th>Key</th><th>値</th><th>説明</th><th>詳細</th></tr></thead>
-      <tbody>${globalRows}</tbody>
-    </table>
-  </details>
-  <details open>
-    <summary>銘柄別設定 (symbol_config) — active ${universe.allowedSymbols.length} / inactive ${universe.inactiveSymbols.length} 銘柄</summary>
-    <p class="muted" style="font-size:12px">
-      inactive (active=0) 銘柄も表示しています。cron / risk gate の評価対象は active=1 のみで、
-      inactive 銘柄は灰色斜体・取消線で区別しています。再有効化は <code>UPDATE symbol_config SET active = 1 WHERE symbol = '...'</code>。
-    </p>
-    <table>
-      <thead><tr><th>銘柄</th><th>状態</th><th>通貨</th><th>1注文あたり上限 (max_notional)</th><th>インバース対 (inverse)</th><th>メモ (notes)</th></tr></thead>
-      <tbody>${symRows}</tbody>
-    </table>
-  </details>`
-}
-
-/**
- * global_config 列のメタ情報 (label + detail)。
- *
- * - `label`: 短い見出し (単位込み)。IT / 汎用英単語 (dry-run / drawdown / spread
- *   等) は英字のまま、日本株固有語 (押し目 / 建玉 / 利食い / 損切り / 騰落率)
- *   のみ日本語化。
- * - `detail`: 株初心者向け advisory。1-3 文、「何をするか」「大小で何が変わるか」
- *   「目安」の順で記述。技術用語を避け具体的な動作で説明。
- *
- * 未登録 key の fallback は em-dash。
- */
-interface ConfigKeyMeta {
-  label: string
-  detail: string
-}
-
-const CONFIG_KEY_META: Record<string, ConfigKeyMeta> = {
-  dry_run: {
-    label: 'dry-run (bool)',
-    detail: 'true にすると実際には注文せず動作確認だけ。false で証券会社へ本当に注文します。テスト中は true、本番のみ false に。',
-  },
-  trading_enabled: {
-    label: 'trading enabled (bool)',
-    detail: 'false にすると全ての注文を拒否します。緊急停止用のスイッチ。止めたい時だけ false に。',
-  },
-  market_hours_check: {
-    label: '場中チェック (bool)',
-    detail: 'true で市場時間外の注文を防ぎます。false は 24 時間発注可 (sandbox 確認用)。',
-  },
-  session_window_gate_enabled: {
-    label: '開場前ゲート (bool)',
-    detail:
-      'true で開場30分前〜引けの窓外は戦略判定を skip (US 09:00–16:00 ET / JP 08:30–15:30 JST、市場ごと)。cron は発火しますが評価しません。false は従来通り常時評価。',
-  },
-  max_order_notional: {
-    label: '1注文上限 (非推奨)',
-    detail: '旧 generic 上限 (通貨別 cap 導入前の互換)。現在は参照されないので触らなくて OK。',
-  },
-  max_order_notional_usd: {
-    label: '1注文上限 (USD)',
-    detail: 'US 株 1 回あたりの発注上限額 (ドル)。大きすぎる注文を防ぐ安全装置。$2000 なら 1 銘柄最大 $2000 まで。',
-  },
-  max_order_notional_jpy: {
-    label: '1注文上限 (JPY)',
-    detail: '日本株 1 回あたりの発注上限額 (円)。同上の円版。¥100000 なら 1 銘柄最大 10 万円まで。',
-  },
-  total_capital_usd: {
-    label: '運用資本 (USD)',
-    detail:
-      'risk-% sizing (stop 距離ベース) の US 株 equity 基準 (ドル)。budget配分% (budget_alloc_pct) 指定銘柄は通貨に関係なく total_capital_jpy 単一プールを使うため、こちらは不要 (USD risk-% 銘柄がある時のみ設定)。',
-  },
-  total_capital_jpy: {
-    label: '運用資本 (JPY / 口座総額)',
-    detail:
-      '口座の運用資本 (円)。budget配分% (budget_alloc_pct) 指定銘柄は通貨に関係なく **この円総額が単一プール基準** (USD 銘柄も USD/JPY で円換算して sizing、#407)。risk-% sizing の日本株 equity 基準も兼ねる。買付余力 pool ゲートの円換算基準でもある (#415)。',
-  },
-  max_portfolio_exposure_pct: {
-    label: 'portfolio exposure 上限率 (比率)',
-    detail: '同時保有の合計上限を「資本 × この率」で決めます。0.6 なら 60%。大きくすると分散度↑、損失時の衝撃↑。',
-  },
-  drawdown_kill_threshold: {
-    label: 'drawdown kill 閾値 (比率、負)',
-    detail: 'その日の損失がこの割合を超えたら、その日は新規売買を止めます。きつく -2% だと早く止まる、緩く -8% だと下げを我慢して継続。',
-  },
-  stale_quote_ms: {
-    label: '気配値鮮度上限 (ms)',
-    detail: '気配値が古すぎる時に判定を止める閾値。900000 = 15 分。短いと厳格、長いと古い気配でも売買。',
-  },
-  gap_reject_pct: {
-    label: 'gap reject 閾値 (比率)',
-    detail: '前日終値からの寄付 gap がこの率を超えた銘柄は買わない。0.03 = 3% 以上の gap で見送り。寄付の高値掴みを防ぐ。',
-  },
-  spread_limit_pct_us: {
-    label: 'spread 上限率 (US、比率)',
-    detail: '買値と売値の差 (spread) がこの率を超えた銘柄は流動性不足で見送り。US は 0.25% 目安。',
-  },
-  spread_limit_pct_jp: {
-    label: 'spread 上限率 (JP、比率)',
-    detail: '買値と売値の差 (spread) がこの率を超えた銘柄は流動性不足で見送り。日本株は 0.6% 目安。',
-  },
-  pullback_default_stop_pct: {
-    label: '損切り幅 (比率、負)',
-    detail: '損切りライン。買値からこの率下がったら売却。-0.04 = -4%。深いと耐えるが大損失リスク、浅いと早く切るが騙し上げで空振り。',
-  },
-  pullback_default_take_profit_pct: {
-    label: '利食い目標 (比率)',
-    detail: '利食い目標。買値からこの率上がったら売却。0.07 = +7%。高いと大きな利益を狙うが取り逃す、低いとコツコツ確定。',
-  },
-  pullback_default_time_stop_days: {
-    label: '最大保有日数 (営業日)',
-    detail: '建玉を保有する最大日数。この日数を超えても利食い/損切りに達しなければ強制売却。10 = 約 2 週間。',
-  },
-  pullback_default_pullback_max: {
-    label: '押し目上限 (比率、負)',
-    detail: '押し目買いを狙う「浅い側」の下落率閾値。-0.03 なら「-3% 以上下げた銘柄を候補に」。緩めると機会↑、騙し↑。',
-  },
-  pullback_default_pullback_min: {
-    label: '押し目下限 (比率、負)',
-    detail: '押し目買いを狙う「深い側」の下落率閾値。-0.06 なら「-6% より深い下げは敬遠」。深すぎる下げは反発せず転換の可能性。',
-  },
-  pullback_default_min_return_50d: {
-    label: '50日最低騰落率 (比率)',
-    detail: '過去 50 日の騰落率がこの値以上の銘柄だけ押し目買い対象。0.08 = +8%。上昇トレンド銘柄を絞るフィルター。',
-  },
-  pullback_default_require_above_sma50: {
-    label: 'SMA50 超必須 (bool)',
-    detail: 'true で 50 日移動平均線より上の銘柄だけ買い対象。上昇トレンドフィルターを厳しくする。',
-  },
-  pullback_default_k_atr: {
-    label: 'ATR 倍率',
-    detail: '損切り幅を ATR (日々の値動き幅) の何倍にするか。2.0 が標準。大きくすると激しい値動き銘柄でも余裕を持って保有、小さいと早めに損切り。',
-  },
-  pullback_default_max_sma50_deviation_pct: {
-    label: '過熱ガード: SMA50 上方乖離上限 (比率)',
-    detail: '株価が 50 日移動平均をこの比率超で上回る過熱局面では押し目買いを見送る。0.6 = +60%。+3x レバ ETF の高値掴み回避。小さいほど厳しく BUY を抑制。',
-  },
-  pullback_default_max_atr_ratio: {
-    label: '過熱ガード: ATR比上限 (倍)',
-    detail: '直近 ATR が baseline (長期平均) のこの倍率を超える高ボラ局面では押し目買いを見送る。1.5 = baseline の 1.5 倍。ボラ・レジーム破綻時の entry を抑制。',
-  },
-  risk_base_per_trade_pct: {
-    label: '基本リスク率 (比率)',
-    detail: '1 回のトレードで失ってよい割合 (対 総資本)。0.004 = 0.4%。大きくすると 1 回あたりの建玉サイズ↑、連敗時の損失↑。',
-  },
-  risk_dd_half_threshold: {
-    label: 'risk half 閾値 (比率、負)',
-    detail: '日次損失がこの率を超えたら 1 回のリスクを半分に減らす。-0.05 = -5%。連敗時の傷を浅く保つ自動ブレーキ。',
-  },
-  risk_dd_halt_threshold: {
-    label: 'risk halt 閾値 (比率、負)',
-    detail: '日次損失がこの率を超えたら 1 回のリスクを 0 に (新規 entry 停止)。-0.10 = -10%。drawdown_kill より前の緊急ブレーキ。',
-  },
-  vix_warning_threshold: {
-    label: 'VIX 警戒閾値',
-    detail: '恐怖指数 (VIX) がこの値を超えたら新規買いの数量を縮小。25 が標準。下げると早めに用心、上げると VIX 高でも普段通り。',
-  },
-  vix_critical_threshold: {
-    label: 'VIX 緊急閾値',
-    detail: '恐怖指数 (VIX) がこの値を超えたら新規買いを全停止 (売却は通常通り)。30 が標準。下げると守り重視、上げると荒れ相場でも買いに行く。',
-  },
-  vix_warning_size_scale: {
-    label: 'VIX 警戒時の建玉縮小率 (比率)',
-    detail: 'VIX 警戒時 (warning ≤ VIX < critical) の発注数量倍率。0.5 = 半分に縮小。1.0 で縮小なし、0 で停止と同義。',
-  },
-}
-
-function formatConfigValue(v: unknown): string {
-  // null placeholder は他ページと同じ em-dash (—) に統一。"null" 文字列は
-  // 運用者が誤って "null" という string 値と混同するリスクがあるので避ける。
-  if (v === null || v === undefined) return '—'
-  if (typeof v === 'boolean') return v ? 'true' : 'false'
-  return String(v)
 }
 
 function chartsBody(args: ChartsBodyArgs): string {
